+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file. This change
 
 ## [2-alpha15-SNAPSHOT]
 - experimental quarto support
+- using the Scicloj Scittle fork for MathBox.cljs support
 
 ## [2-alpha14]  - 2023-02-21
 - table view support for datasets